feat(recruitment): enterprise resume scoring — ATS engine, bulk upload pipeline, MCQ bank, resume insights
- ATS scoring engine: 5-dimension scoring (sections/25, keywords/35, contact/15,
  quantified/15, parse quality/10) — real scores, no fake inflation
- Bulk resume upload: 6 new routes + async per-resume AI+ATS scoring via setImmediate
  (fire-and-forget), BulkResumeUpload/BulkResumeItem Prisma models with org indexes
- Expanded MCQ fallback bank: 30 questions (10 × INTELLIGENCE/INTEGRITY/ENERGY),
  Fisher-Yates shuffle + per-candidate random draw — every candidate sees unique set
- Cloud resume URL support: fetch() download to temp buffer, local path fallback,
  cleanup after processing — fixes null scores for cloud-hosted resumes
- Resume intelligence panel: strengths/gaps cards, keyword match chips (found/missing),
  collapsible extracted text, parse method badge, ATS breakdown progress bars
- Fallback question warning banner: shown when AI MCQ generation was unavailable
- New Prisma fields on PublicApplication: resumeText, matchedKeywords, missingKeywords,
  atsScore, atsScoreData, usedFallbackQuestions
- Removed hardcoded dev-ai-key fallback + score 1.2× inflation hack
- ESM static imports (path, fs, crypto) replacing inline require() calls

+++ backend/src/modules/payroll/payroll.validation.ts
@@ -89,7 +89,7 @@ export const amendPayrollRecordSchema = z.object({
   esiEmployee: z.number().min(0).optional(),
   professionalTax: z.number().min(0).optional(),
   tds: z.number().min(0).optional(),
-  lopDays: z.number().int().min(0).optional(),
+  lopDays: z.number().min(0).optional(), // Decimal — half-days (0.5) are valid
   lopDeduction: z.number().min(0).optional(),
   reason: z.string().min(1, 'Amendment reason is required'),
 });
